Rename entrypoint to esm.ts
