+++ packages/upscalerjs/tsconfig.esm.json
@@ -4,6 +4,6 @@
     "target": "ESNext",
     "module": "ESNext",
   },
-  "include": ["src/index.ts"],
+  "include": ["src/esm.ts"],
   "exclude": ["node_modules", "**/*.test.ts"]
 }
