+++ test/integration/browser/upscale.ts
@@ -8,7 +8,7 @@ import Upscaler from 'upscaler';
 import { BrowserTestRunner } from '../utils/BrowserTestRunner';
 import path from 'path';
 import { MODELS_DIR } from '../../../scripts/package-scripts/utils/constants';
-import { MultiArgStringProgress, MultiArgTensorProgress } from '../../../packages/upscalerjs/src';
+import { MultiArgStringProgress, MultiArgTensorProgress } from '../../../packages/upscalerjs/src/esm';
 
 const PIXEL_UPSAMPLER_DIR = path.resolve(MODELS_DIR, 'pixel-upsampler/test/__fixtures__');
